// Neptune.Web/src/app/shared/generated/model/treatment-bmp-assessment-detail-dto.ts
export class TreatmentBMPAssessmentDetailDto { 
    TreatmentBMPAssessmentID?: number;
    FieldVisitID?: number;
    TreatmentBMPID?: number;
    TreatmentBMPTypeID?: number;
    TreatmentBMPAssessmentTypeID?: number;
    TreatmentBMPAssessmentTypeDisplayName?: string | null;
    IsAssessmentComplete?: boolean;
    AssessmentScore?: number | null;
    Notes?: string | null;
    Observations?: Array<TreatmentBMPObservationDto> | null;
    ObservationTypes?: Array<TreatmentBMPAssessmentObservationTypeForFormDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}