// Neptune.Web/src/app/shared/generated/model/treatment-bmp-assessment-observation-type-for-form-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPAssessmentObservationTypeForFormDto { 
    TreatmentBMPAssessmentObservationTypeID?: number;
    TreatmentBMPAssessmentObservationTypeName?: string | null;
    ObservationTypeSpecificationID?: number;
    ObservationTypeCollectionMethodName?: string | null;
    TreatmentBMPAssessmentObservationTypeSchema?: string | null;
    SortOrder?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPAssessmentObservationTypeForFormDtoForm { 
    TreatmentBMPAssessmentObservationTypeID?: FormControl<number>;
    TreatmentBMPAssessmentObservationTypeName?: FormControl<string>;
    ObservationTypeSpecificationID?: FormControl<number>;
    ObservationTypeCollectionMethodName?: FormControl<string>;
    TreatmentBMPAssessmentObservationTypeSchema?: FormControl<string>;
    SortOrder?: FormControl<number>;
}

export class TreatmentBMPAssessmentObservationTypeForFormDtoFormControls { 
    public static TreatmentBMPAssessmentObservationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPAssessmentObservationTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ObservationTypeSpecificationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ObservationTypeCollectionMethodName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPAssessmentObservationTypeSchema = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}