// Neptune.Web/src/app/shared/generated/model/treatment-bmp-assessment-score-detail-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TreatmentBMPAssessmentObservationTypeForScoringDto } from './treatment-bmp-assessment-observation-type-for-scoring-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPAssessmentScoreDetailDto { 
    TreatmentBMPAssessmentID?: number;
    IsAssessmentComplete?: boolean;
    IsBenchmarkAndThresholdsComplete?: boolean;
    AssessmentScore?: string | null;
    OverrideScore?: boolean;
    ObservationTypes?: Array<TreatmentBMPAssessmentObservationTypeForScoringDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPAssessmentScoreDetailDtoForm { 
    TreatmentBMPAssessmentID?: FormControl<number>;
    IsAssessmentComplete?: FormControl<boolean>;
    IsBenchmarkAndThresholdsComplete?: FormControl<boolean>;
    AssessmentScore?: FormControl<string>;
    OverrideScore?: FormControl<boolean>;
    ObservationTypes?: FormControl<Array<TreatmentBMPAssessmentObservationTypeForScoringDto>>;
}

export class TreatmentBMPAssessmentScoreDetailDtoFormControls { 
    public static TreatmentBMPAssessmentID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsAssessmentComplete = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsBenchmarkAndThresholdsComplete = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AssessmentScore = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OverrideScore = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ObservationTypes = (value: FormControlState<Array<TreatmentBMPAssessmentObservationTypeForScoringDto>> | Array<TreatmentBMPAssessmentObservationTypeForScoringDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<TreatmentBMPAssessmentObservationTypeForScoringDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}